union_product_marker_webserver/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
pandas==2.1.4
openpyxl==3.1.2
Pillow>=10.0.0 